safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-size-sync.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../network-test.js"></script>
<script>
function test()
{
    InspectorTest.makeSimpleXHR("GET", "resources/resource.php?type=image&size=500&gzip=true&chunked=true", false, step2);

    function step2()
    {
        var request1 = WebInspector.panel("network").requests[WebInspector.panel("network").requests.length - 1];
        InspectorTest.addResult(request1.url);
        InspectorTest.addResult("resourceSize: " + request1.resourceSize);
        InspectorTest.addResult("transferSize: " + request1.transferSize);

        InspectorTest.completeTest();
    }
}
</script>
</head>
<body onload="runTest()">
<p>Tests network resource size reporting for synchronous requests.</p>
</body>
</html>
