safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-size-sync-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests network resource size reporting for synchronous requests.

http://127.0.0.1:8000/inspector/network/resources/resource.php?type=image&size=500&gzip=true&chunked=true
resourceSize: 500
transferSize: 937
